+++ src/check.rs
@@ -117,34 +117,11 @@ impl<'a> VirtualEnv<'a> {
                 self.handle_args_outputs(0, *count)?
             }
             Instr::PushFunc(f) => self.function_stack.push(Cow::Borrowed(f)),
-            &Instr::Switch { count, .. } => {
-                let mut funcs = Vec::with_capacity(count);
+            &Instr::Switch { count, sig, .. } => {
                 for _ in 0..count {
-                    funcs.push(self.pop_func()?);
+                    self.pop_func()?;
                 }
-                match self.pop()? {
-                    BasicValue::Num(i) if i >= 0.0 && i.fract() == 0.0 && (i as usize) < count => {
-                        let i = i as usize;
-                        self.handle_sig(funcs[i].signature())?
-                    }
-                    _ => {
-                        if let Some(win) = funcs
-                            .windows(2)
-                            .find(|win| !win[0].signature().is_compatible_with(win[1].signature()))
-                        {
-                            return Err(format!(
-                                "switch's functions have incompatible signatures {} and {}",
-                                win[0].signature(),
-                                win[1].signature()
-                            ));
-                        }
-                        let sig = funcs
-                            .iter()
-                            .map(|f| f.signature())
-                            .fold(Signature::new(0, 0), |a, b| a.max_with(b));
-                        self.handle_sig(sig)?
-                    }
-                }
+                self.handle_args_outputs(sig.args + 1, sig.outputs)?;
             }
             Instr::Dynamic(f) => self.handle_sig(f.signature)?,
             Instr::DropTemp { .. } => {}