+++ changelog.md
@@ -6,6 +6,7 @@ Uiua is not yet stable.
 This version is not yet released.
 If you are reading this on the website, then these changes are live here.
 ### Language
+- Switch functions are now less strict about branch signature compatibility
 - [`under` `⍜`](https://uiua.org/docs/under) now works with [`absolut value` `⌵`](https://uiua.org/docs/absolute)
 ### Interpreter
 - Lots of bug and crash fixes